make file_watcher a standalone module
