+++ lib/cortex/controller.ex
@@ -20,50 +20,21 @@ defmodule Cortex.Controller do
   ##########################################
 
   def init(_) do
-    dirs = ["lib/", "test/", "apps/"] |> Enum.filter(&File.dir?/1)
-    {:ok, watcher_pid} = FileSystem.start_link(dirs: dirs)
-    FileSystem.subscribe(watcher_pid)
-
     pipeline =
       case Mix.env do
         :dev -> [Reloader]
         :test -> [Reloader, TestRunner]
       end
 
-    {:ok, %{watcher_pid: watcher_pid, pipeline: pipeline}}
+    {:ok, %{pipeline: pipeline}}
   end
 
-  def handle_info({:file_event, watcher_pid, {path, _events}}, %{watcher_pid: watcher_pid}=state) do
-    run_pipeline(state.pipeline, file_type(path), path)
+  def handle_cast({:file_changed, type, path}, %{pipeline: pipeline} = state) do
+    run_pipeline(pipeline, type, path)
+
     {:noreply, state}
   end
 
-  def handle_info({:file_event, watcher_pid, :stop}, %{watcher_pid: watcher_pid}=state) do
-    Logger.info "File watcher stopped."
-    {:noreply, state}
-  end
-
-  def handle_info(data, state) do
-    Logger.info "Get unexcepted message #{inspect data}, ignore..."
-    {:noreply, state}
-  end
-
-
-  def file_type(path) do
-    is_elixir_file? =
-      Regex.match?(~r/\/(\w|_)+\.exs?/, path)
-
-    cond do
-      is_elixir_file? and Regex.match?(~r/lib\//, path) ->
-        :lib
-
-      is_elixir_file? and Regex.match?(~r/test\//, path) ->
-        :test
-
-      true ->
-        :unknown
-    end
-  end
 
   ##########################################
   # Stage Module
